:recycle: refactor register.js 리팩터링 및 파일 분리
