+++ client/src/components/ReservationComponents/useReservation.js
@@ -1,43 +1,68 @@
 import axios from 'axios';
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { useNavigate, useParams } from 'react-router-dom';
+import { useRecoilState } from 'recoil';
 import header from '../../utils/header';
 
 import {
   reservationStartDate,
   reservationEndDate,
   reservationMaxCapacity,
-  PlaceIDState,
 } from '../../atoms';
+
+import { onClickPaymentButton } from '../../utils/payment';
 
 const useReservation = charge => {
   const [startDate, setStartDate] = useRecoilState(reservationStartDate);
   const [endDate, setEndDate] = useRecoilState(reservationEndDate);
   const [capacity, setCapacity] = useRecoilState(reservationMaxCapacity);
-  const placeId = useRecoilValue(PlaceIDState);
   const [modalOpen, setModalOpen] = useState(false);
+  const [reserveId, setReserveId] = useState(null);
+  // const [returnInformation, setReturnInformation] = useState(null);
+
+  const { id } = useParams();
 
   const navigator = useNavigate();
 
-  const timeDiff = new Date(endDate).getTime() - new Date(startDate).getTime();
-  const reservedTimeRange = timeDiff / (1000 * 60 * 60);
+  const getReservationInformation = () => {
+    const timeDiff =
+      new Date(endDate).getTime() - new Date(startDate).getTime();
+    const reservedTimeRange = timeDiff / (1000 * 60 * 60);
 
-  const chargePerHour = charge;
-  const chargePerHourString = new Intl.NumberFormat('ko-KR', {
-    currency: 'KRW',
-    // style: 'currency',
-    currencyDisplay: 'code',
-  }).format(charge);
+    const chargePerHourString = new Intl.NumberFormat('ko-KR').format(charge);
 
-  const totalCharge = reservedTimeRange * chargePerHour;
-  const totalChargeString = new Intl.NumberFormat('ko-KR').format(totalCharge);
+    const totalCharge = reservedTimeRange * charge;
+    const totalChargeString = new Intl.NumberFormat('ko-KR').format(
+      totalCharge,
+    );
+
+    const chargePerHour = `${chargePerHourString}원`;
+    const totalReservedTimeRange = `총 ${reservedTimeRange}시간`;
+    const totalCharges = `총 ${totalChargeString}원`;
+    const chargeMultiplyTime = `총 ${chargePerHourString}원 x ${reservedTimeRange}시간`;
+    return {
+      chargePerHour,
+      totalReservedTimeRange,
+      totalCharges,
+      chargeMultiplyTime,
+    };
+  };
 
   const onClickPaymentKaKaoButton = async () => {
-    // const response = await axios.get(`/place/reserve/?/payment`);
-    // const paymentUrl = response.data.data;
-    // onClickPaymentButton(paymentUrl);
-    setModalOpen(false);
+    try {
+      const response = await axios.get(
+        `/place/reserve/${reserveId}/payment`,
+        header,
+      );
+      const paymentUrl = response.data.data;
+      onClickPaymentButton(paymentUrl);
+      setModalOpen(false);
+
+      const result = await axios.get(`/api/order/completed?pg_token=`, header);
+      console.log(result);
+    } catch (error) {
+      console.log('Error', error);
+    }
   };
 
   const IsPayment = {
@@ -69,12 +94,12 @@ const useReservation = charge => {
 
     try {
       const response = await axios.post(
-        `/place/${placeId}/reserve`,
-        // JSON.stringify(reservationInformation),
+        `/place/${id}/reserve`,
         reservationInformation,
         header,
       );
-      console.log(response.data);
+      console.log(response);
+      setReserveId(response.data);
       setStartDate(false);
       setEndDate(false);
     } catch (err) {
@@ -83,7 +108,6 @@ const useReservation = charge => {
   };
 
   return {
-    chargePerHourString,
     startDate,
     setStartDate,
     endDate,
@@ -93,9 +117,8 @@ const useReservation = charge => {
     handleSubmit,
     modalOpen,
     IsPayment,
-    // onClickPaymentButton,
-    reservedTimeRange,
-    totalChargeString,
+    onClickPaymentButton,
+    getReservationInformation,
   };
 };
